Renamed files Seedfinder* to SeedFinder* and necessary includes
