+++ Core/include/Acts/Seeding/SeedFinder.ipp
