+++ Plugins/Cuda/include/Acts/Plugins/Cuda/Seeding2/Details/FindTriplets.hpp
@@ -67,7 +67,7 @@ namespace Details {
 ///            @c Acts::SeedFinderConfig
 /// @param[in] pT2perRadius Configuration parameter from
 ///            @c Acts::SeedFinderConfig
-/// @param[in] impactMax Configuration parameter from @c Acts::SeedfinderConfig
+/// @param[in] impactMax Configuration parameter from @c Acts::SeedFinderConfig
 /// @return A 2-D structure holding the parameters of the identified triplets
 ///         for each middle spacepoint
 ///
