+++ Tests/UnitTests/Core/Seeding/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(ActsUnitTestSeedfinder SeedfinderTest.cpp)
-target_link_libraries(ActsUnitTestSeedfinder PRIVATE ActsCore Boost::boost)
+add_executable(ActsUnitTestSeedFinder SeedFinderTest.cpp)
+target_link_libraries(ActsUnitTestSeedFinder PRIVATE ActsCore Boost::boost)
 
 add_unittest(EstimateTrackParamsFromSeedTest EstimateTrackParamsFromSeedTest.cpp)
