+++ Tests/UnitTests/Plugins/Cuda/Seeding/CMakeLists.txt
@@ -1,3 +1,3 @@
-add_executable(ActsUnitTestSeedfinderCuda SeedfinderCudaTest.cpp)
-target_link_libraries(ActsUnitTestSeedfinderCuda PRIVATE ${unittest_extra_libraries} Boost::boost)
+add_executable(ActsUnitTestSeedFinderCuda SeedFinderCudaTest.cpp)
+target_link_libraries(ActsUnitTestSeedFinderCuda PRIVATE ${unittest_extra_libraries} Boost::boost)
