+++ Tests/UnitTests/Plugins/Sycl/Seeding/CMakeLists.txt
@@ -6,12 +6,12 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-add_executable(ActsUnitTestSeedfinderSycl
-  SeedfinderSyclTest.cpp
+add_executable(ActsUnitTestSeedFinderSycl
+  SeedFinderSyclTest.cpp
   CommandLineArguments.h
   CommandLineArguments.cpp
   ATLASCuts.hpp
   SpacePoint.hpp)
-target_link_libraries(ActsUnitTestSeedfinderSycl
+target_link_libraries(ActsUnitTestSeedFinderSycl
   PRIVATE Boost::program_options ActsCore ActsPluginSycl
           vecmem::core vecmem::sycl)
